utils file added with string utils
